chore: move old files to a legacy folder
