+++ src/features/errorModal/ErrorModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Modal from 'react-modal'
 
-import cstyles from '../../components/Common.module.css'
+import cstyles from '../../legacy/components/Common.module.css'
 import { useAppDispatch, useAppSelector } from '../../redux/hooks'
 import { closeErrorModal } from './errorModalSlice'
