+++ src/features/pastelID/PastelID.tsx
@@ -5,11 +5,11 @@ import {
 import React, { useEffect, useState } from 'react'
 import { v4 as uid } from 'uuid'
 
-import cstyles from '../../components/Common.module.css'
-import List from '../../components/List'
-import ListItem from '../../components/ListItem'
-import LoadingOverlay from '../../components/LoadingOverlay'
-import Select from '../../components/Select'
+import LoadingOverlay from '../..//legacy/components/LoadingOverlay'
+import cstyles from '../../legacy/components/Common.module.css'
+import List from '../../legacy/components/List'
+import ListItem from '../../legacy/components/ListItem'
+import Select from '../../legacy/components/Select'
 import { useAppDispatch, useAppSelector } from '../../redux/hooks'
 import { openErrorModal } from '../errorModal'
 import styles from './PastelID.module.css'
