+++ src/index.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 
-import Root from './containers/Root'
+import Root from './legacy/containers/Root'
 import store from './redux/store'
 
 const application = (
