+++ src/legacy/assets/img/arrow_up_dark.png
