+++ src/legacy/assets/img/pastel-logo.png
