+++ src/legacy/assets/img/pastel-logo2.png
