+++ src/legacy/companion.ts
