+++ src/legacy/components/ErrorModal.tsx
