+++ src/legacy/containers/Root.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter as Router } from 'react-router-dom'
 
-import { ErrorModal as ReduxErrorModal } from '../features/errorModal'
+import { ErrorModal as ReduxErrorModal } from '../../features/errorModal'
 import Routes from '../Routes'
 
 const Root = (): JSX.Element => (
