+++ src/legacy/utils/uris.test.js
